rename: quiz landing(#59)
