+++ strawberry/src/pages/quizLanding/QuizLandingPage.tsx
@@ -1,6 +1,7 @@
+import { Wrapper } from "../../core/design_system";
+
 import QuizBanner from "./components/quizBanner";
 import QuizInfo from "./components/quizInfo";
-import { Wrapper } from "../../core/design_system";
 import QuizOpen from "./components/quizOpen";
 import QuizReward from "./components/quizReward";
 import QuizNotice from "./components/quizNotice";
