+++ strawberry/src/pages/quizLanding/components/quizBanner/QuizBanner.tsx
