+++ strawberry/src/pages/quizLanding/components/quizBanner/QuizBannerContent.tsx
