+++ strawberry/src/pages/quizLanding/components/quizBanner/QuizBannerTimer.tsx
