+++ strawberry/src/pages/quizLanding/components/quizBanner/TimerSection.tsx
