+++ strawberry/src/pages/quizLanding/components/quizBanner/index.tsx
