+++ strawberry/src/pages/quizLanding/components/quizInfo/QuizInfo.tsx
