+++ strawberry/src/pages/quizLanding/components/quizInfo/index.tsx
