+++ strawberry/src/pages/quizLanding/components/quizOpen/QuizOpen.tsx
