+++ strawberry/src/pages/quizLanding/components/quizOpen/QuizOpenStep.tsx
