+++ strawberry/src/pages/quizLanding/components/quizOpen/index.tsx
