+++ strawberry/src/pages/quizLanding/components/quizReward/QuizReward.tsx
