+++ strawberry/src/pages/quizLanding/components/quizReward/QuizRewardItem.tsx
