+++ strawberry/src/pages/quizLanding/components/quizReward/index.tsx
